backend/requirements.txt
openai
python-dotenv
langchain
fastapi 
uvicorn